+++ src/ralph/ui/forms/reports.py
@@ -6,6 +6,7 @@ from __future__ import unicode_literals
 
 from django import forms
 from ralph.ui.widgets import DateWidget
+from ralph.ui.forms import all_ventures
 
 
 class DevicesChoiceReportForm(forms.Form):
@@ -55,3 +56,11 @@ class DeprecationRangeReportForm(forms.Form):
 class WarrantyRangeReportForm(forms.Form):
     w_start = forms.DateField(widget=DateWidget, label='Start date')
     w_end = forms.DateField(widget=DateWidget, label='End date')
+
+
+class ReportVentureCost(forms.Form):
+    venture = forms.ChoiceField(label="Venture")
+
+    def __init__(self, *args, **kwargs):
+        super(ReportVentureCost, self).__init__(*args, **kwargs)
+        self.fields['venture'].choices = all_ventures()