+++ src/ralph/ui/tests/test_deployment.py
@@ -19,7 +19,7 @@ from ralph.discovery.models import (
 )
 from ralph.business.models import Venture, VentureRole
 from ralph.deployment.models import Preboot, Deployment
-from ralph.ui.forms import (
+from ralph.ui.forms.deployment import (
     _validate_cols_count,
     _validate_cols_not_empty,
     _validate_mac,
@@ -56,11 +56,9 @@ class BulkDeploymentTest(TestCase):
             _validate_mac('sproing', ['deadbeefcafe'], 0)
 
     def test_valdate_management_ip(self):
-        _validate_management_ip('127.0.0.1', [], 0)
+        _validate_management_ip('127.0.0.1', 0)
         with self.assertRaises(forms.ValidationError):
-            _validate_management_ip('127.0.0.0.1', [], 0)
-        with self.assertRaises(forms.ValidationError):
-            _validate_management_ip('127.0.0.1', ['127.0.0.1'], 0)
+            _validate_management_ip('127.0.0.0.1', 0)
 
     def test_validate_network_name(self):
         with self.assertRaises(forms.ValidationError):
@@ -159,6 +157,11 @@ class BulkDeploymentTest(TestCase):
 
     def test_validate_ip_owner(self):
         _validate_ip_owner('127.0.0.1', 'deadbeefcafe', 0)
+        another_device = Device.create(
+            ethernets=[('', 'deadbeefcafd', 0)],
+            model_name='splat',
+            model_type=DeviceType.unknown,
+        )
         ip = IPAddress(address='127.0.0.1')
         ip.save()
         with self.assertRaises(forms.ValidationError):
@@ -171,8 +174,9 @@ class BulkDeploymentTest(TestCase):
         ip.device.save()
         ip.save()
         _validate_ip_owner('127.0.0.1', 'deadbeefcafe', 0)
-        ip.device = None
+        ip.device = another_device
         ip.save()
         with self.assertRaises(forms.ValidationError):
             _validate_ip_owner('127.0.0.1', 'deadbeefcafe', 0)
+